normalization improvements

+++ genbank_dump_script.sh
@@ -1,4 +1,4 @@
-# Run this script with sudo 
+# Run this script with sudo or add it to the cron job
 # Genbank dump runs bimonthly
 #
 
@@ -14,9 +14,7 @@
 DB_USER=zoophyadmin
 DB_HOST='zodo.asu.edu'
 DB_NAME='GenBankViruses'_`date +%b%Y`
-#DB_NAME='GenBankViruses_Dec2017'
 CREATE_DB="create database $DB_NAME;"
-SMALL_DB=''
 JDBC_DB_URL='jdbc:postgresql://zodo.asu.edu/GenBankViruses_UI'
 UI_DB='GenBankViruses_UI'
 # give postgres password here
@@ -28,12 +26,11 @@ psql -h $DB_HOST -U $DB_USER template1 -c "create database \"$DB_NAME\";"
 echo "created database ..!!"
 
 # Taking backup of old UI db
-pg_dump -h $DB_HOST -U $DB_USER -F c GenBankViruses_UI > /home/zoophy/old_versions/GenBankViruses_UI_DB_before_`date +%b%Y`.backup
-
 echo "taking backup of old ui db ..!"
+pg_dump -h $DB_HOST -U $DB_USER -F c $UI_DB > /home/zoophy/old_versions/GenBankViruses_UI_DB_before_`date +%b%Y`.backup
 
 # Fetch latest code of zoophy-genbankfactory
-echo "Fetching latest code "
+echo "Fetching latest code for zoophy-genbankfactory"
 
 cd /home/zoophy/genbank/zoophy-genbankfactory ; git pull
 
@@ -48,9 +45,9 @@ echo "Done fetching latest code ..!"
 echo "Updating config file .!"   
 
 
-sed -i "s/^\(DB\.Big\.Name\s*=\s*\).*\$/\1$DB_NAME/" src/GenBankFactory.local.properties 
+sed -i "s/^\(DB\.Big\.Name\s*=\s*\).*\$/\1$DB_NAME/" GenBankFactory.local.properties 
 
-#sed -i "s/^\(DB\.Small\.Name\s*=\s*\).*\$/\1$DB_NAME/" src/GenBankFactory.local.properties 
+sed -i "s/^\(DB\.Small\.Name\s*=\s*\).*\$/\1$UI_DB/" GenBankFactory.local.properties 
 
 echo "Done updating the config file"
 
@@ -68,8 +65,10 @@ echo "command"
 java -Xms4G -Xmx8G -jar target/zoophy-genbank-factory-*-jar-with-dependencies.jar dump create -f gbvrl 
 
 #############################################################################################################################################################
+# we need to run the following part only when build and dump are successful
+# it might be better to run these command manually as individual paths might change
 
-# backup the old ui_index_main in /home/zoophy/old_versions/
+# backup the old ui_index_main 
 echo "backup the old ui_index in /home/zoophy/old_versions/ "
 cp -r /home/zoophy/indices/ui_index_main /home/zoophy/old_versions/ui_index_before_`date +%b%Y`
 
@@ -81,29 +80,32 @@ echo "replace old lucene index with newly created "
 cp -r /home/zoophy/indices/small_index_main/* /home/zoophy/indices/ui_index_main 
 
 
-# Edit configuration file to update the new db name in zoophy-services config file
-echo "updating zoophy services config "
-sed -i "s|^\(spring\.datasource\.url\s*=\s*\).*\$|\1$JDBC_DB_URL|" /home/zoophy/zoophy-services/config/application.properties
+if ["$1" != "" -a $1 == "restart_services"];
+    # Edit configuration file to update the new db name in zoophy-services config file
+    echo "updating zoophy services config "
+    sed -i "s|^\(spring\.datasource\.url\s*=\s*\).*\$|\1$JDBC_DB_URL|" /home/zoophy/zoophy-services/config/application.properties
 
-# Build & restart the services
-echo "Building zoophy-services.."
-cd /home/zoophy/zoophy-services ; git pull ; ./build.sh
+    # Build & restart the services
+    echo "Building zoophy-services.."
+    cd /home/zoophy/zoophy-services ; git pull ; ./build.sh
 
-# deploy new zoophy services 
-# if jar name is changes then edit jar name here
-# kill the old zoophy services
-process_id=`/bin/ps -fu $USER| grep "java -jar target/zoophy-rest-service*" | grep -v "grep" | awk '{print $2}'`
-echo "kiiling the old process "$process_id
-kill -9 $process_id
+    # deploy new zoophy services 
+    # if jar name is changed then edit jar name here
+    # kill the old zoophy services
+    process_id=`/bin/ps -fu $USER| grep "java -jar target/zoophy-rest-service*" | grep -v "grep" | awk '{print $2}'`
+    echo "killing the old zoophy rest service process "$process_id
+    kill -9 $process_id
 
-# start spring boot app
-echo "starting zoophy rest services..! "
-java -jar target/zoophy-rest-service*.jar > zoophy_services_`date "+%Y-%m-%d_%H:%M:%S"`.log 2>&1 & 
+    # start spring boot app
+    echo "starting zoophy rest services..! "
+    java -jar target/zoophy-rest-service*.jar > zoophy_services_`date "+%Y-%m-%d_%H:%M:%S"`.log 2>&1 & 
 
-# Setup the zoophy-ui
-echo "fetching zoophy ui"
-cd /home/zoophy/zoophy-ui ; git pull
-# no config change unless services port is been changed
-# restart app
-pm2 restart zoophy
-echo "End of Script"
+    # Setup the zoophy-ui
+    echo "fetching zoophy ui"
+    cd /home/zoophy/zoophy-ui ; git pull
+    # no config change unless services port is been changed
+    # restart app
+    pm2 restart zoophy
+    echo "End of Script"
+
+echo "replace old lucene index with newly created "
